+++ spec/Processors/WriteOutProcessorSpec.php
@@ -5,7 +5,7 @@ namespace spec\Rorschach\Processors;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
 use Rorschach\Config;
-use Rorschach\Step;
+use Rorschach\Checkpoint;
 
 class WriteOutProcessorSpec extends ObjectBehavior
 {
@@ -20,7 +20,7 @@ class WriteOutProcessorSpec extends ObjectBehavior
         $config->getWriteOut()->willReturn($dir);
         $this->beConstructedWith($config);
 
-        $this->getWrappedObject()->process(new Step('Test', '/path'), 'data');
+        $this->getWrappedObject()->process(new Checkpoint('Test', '/path'), 'data');
         expect(\file_exists($dir . '/Test.png'))->toBe(true);
         expect(\file_get_contents($dir . '/Test.png'))->toBe('data');
         `rm -rf $dir`;
@@ -37,7 +37,7 @@ class WriteOutProcessorSpec extends ObjectBehavior
         $config->getWriteOut()->willReturn($dir);
         $this->beConstructedWith($config);
 
-        $this->getWrappedObject()->process(new Step('Name with / funny + chars &', '/path'), 'data');
+        $this->getWrappedObject()->process(new Checkpoint('Name with / funny + chars &', '/path'), 'data');
         expect(\file_exists($dir . '/Name+with+%2F+funny+%2B+chars+%26.png'))->toBe(true);
         expect(\file_get_contents($dir . '/Name+with+%2F+funny+%2B+chars+%26.png'))->toBe('data');
         `rm -rf $dir`;